+++ src/resources/event/event.controllers.js
@@ -1,4 +1,4 @@
-import { crudControllers } from './eventCurd'
+import { crudControllers } from '../../utils/eventCurd'
 import { Event } from './event.model'
 
 export default crudControllers(Event)
